+++ packages/wallet-engine/src/contracts.ts
@@ -1,40 +1,46 @@
 import type { Signer } from 'ethers';
 import {
   ApplicationEnum,
+  Deployment,
+  ChainId,
   getContractAddress as getSCAddress,
-  createERC20,
-  getTokenConfig as fpGetTokenConfig,
+  createERC20MockToken,
   Freeport__factory,
-  TestERC20__factory,
-  SimpleAuction__factory,
-  Auction__factory,
-  CollectionFactory__factory,
+  ERC20MockToken__factory,
   Marketplace__factory,
-} from '@cere/freeport-sdk';
+  FreeportCollection__factory,
+  FiatGateway__factory,
+  FToken__factory,
+  ERC20MockToken,
+} from '@cere/freeport-sc-sdk';
 
 import { CERE_TOKEN_ADDRESS } from './constants';
 
-export type { TokenConfig } from '@cere/freeport-sdk';
+export type TokenConfig = {
+  symbol: string;
+  decimals: number;
+};
 
+export type ERC20Contract = ERC20MockToken;
 export enum ContractName {
   Freeport = 'Freeport',
-  ERC20 = 'ERC20',
-  SimpleAuction = 'SimpleAuction',
-  Auction = 'Auction',
+  FToken = 'FToken',
   Marketplace = 'Marketplace',
-  CollectionFactory = 'CollectionFactory',
+  FreeportCollection = 'FreeportCollection',
+  FiatGateway = 'FiatGateway',
+  ERC20Token = 'ERC20MockToken',
   CereToken = 'CereToken',
 }
 
 const applications = [ApplicationEnum.DAVINCI, ApplicationEnum.LIVEONE];
 const contractInterfaceFactoryMap = {
   [ContractName.Freeport]: () => Freeport__factory.createInterface(),
-  [ContractName.ERC20]: () => TestERC20__factory.createInterface(),
-  [ContractName.SimpleAuction]: () => SimpleAuction__factory.createInterface(),
-  [ContractName.Auction]: () => Auction__factory.createInterface(),
+  [ContractName.ERC20Token]: () => ERC20MockToken__factory.createInterface(),
   [ContractName.Marketplace]: () => Marketplace__factory.createInterface(),
-  [ContractName.CollectionFactory]: () => CollectionFactory__factory.createInterface(),
-  [ContractName.CereToken]: () => TestERC20__factory.createInterface(),
+  [ContractName.FreeportCollection]: () => FreeportCollection__factory.createInterface(),
+  [ContractName.CereToken]: () => ERC20MockToken__factory.createInterface(),
+  [ContractName.FiatGateway]: () => FiatGateway__factory.createInterface(),
+  [ContractName.FToken]: () => FToken__factory.createInterface(),
 };
 
 export const getContractInterface = (contractName: ContractName) => {
@@ -53,8 +59,8 @@ export const getContractAddress = (
   return getSCAddress({
     application,
     contractName,
-    chainId: parseInt(networkId, 16),
-    deployment: process.env.REACT_APP_ENV,
+    chainId: parseInt(networkId, 16) as ChainId,
+    deployment: process.env.REACT_APP_ENV as Deployment,
   });
 };
 
@@ -77,16 +83,20 @@ export const getContractNameByAddress = (address: string, networkId: string) => 
   return contractNames.find((name) => getAllContractAddresses(name, networkId).includes(loweredAddress));
 };
 
-export const createUSDCContract = (signer: Signer, networkId: string, application?: ApplicationEnum) =>
-  createERC20({
-    signer,
-    contractAddress: getContractAddress(ContractName.ERC20, networkId, application),
-  });
-
 export const createERC20Contract = (signer: Signer, address: string) =>
-  createERC20({
+  createERC20MockToken({
     signer,
     contractAddress: address,
   });
 
-export const getTokenConfig = () => fpGetTokenConfig(process.env.REACT_APP_ENV);
+export const getERC20TokenConfig = async (signer: Signer): Promise<TokenConfig> => {
+  const chainId = await signer.getChainId();
+  const address = getContractAddress(ContractName.ERC20Token, chainId.toString(16));
+  const contract = createERC20Contract(signer, address);
+  const decimals = await contract.decimals();
+
+  return {
+    symbol: 'USDC', // Keep it hardcoded for better UX
+    decimals,
+  };
+};