+++ src/stores/AssetStore/UsdcToken.ts
@@ -1,81 +1,16 @@
-import { utils } from 'ethers';
-import { makeAutoObservable } from 'mobx';
-import { fromResource } from 'mobx-utils';
-import { createUSDCContract, getTokenConfig, TokenConfig } from '@cere-wallet/wallet-engine';
+import { ContractName, getContractAddress } from '@cere-wallet/wallet-engine';
 
-import { TransferableAsset, ReadyWallet } from './types';
+import { ReadyWallet } from './types';
+import { Erc20Token } from './Erc20Token';
 
-const createBalanceResource = ({ provider, network, account }: ReadyWallet, { decimals }: TokenConfig) => {
-  let currentListener: () => {};
-
-  const erc20 = createUSDCContract(provider.getSigner(), network.chainId);
-  const receiveFilter = erc20.filters.Transfer(null, account.address);
-  const sendFilter = erc20.filters.Transfer(account.address);
-
-  return fromResource<number>(
-    (sink) => {
-      currentListener = async () => {
-        const balance = await erc20.balanceOf(account.address);
-
-        sink(balance.div(10 ** decimals).toNumber());
-      };
-
-      /**
-       * Prefetch erc20 token balance
-       */
-      currentListener();
-
-      /**
-       * Update erc20 token balance on each transfer event
-       */
-      provider!.on(receiveFilter, currentListener);
-      provider!.on(sendFilter, currentListener);
-    },
-    () => {
-      provider!.off(receiveFilter, currentListener);
-      provider!.off(sendFilter, currentListener);
-    },
-  );
-};
-
-export class UsdcToken implements TransferableAsset {
-  private tokenConfig = getTokenConfig();
-  private balanceResource = createBalanceResource(this.wallet, this.tokenConfig);
-  public id: string = 'matic-network';
-
-  constructor(private wallet: ReadyWallet) {
-    makeAutoObservable(this);
-  }
-
-  get decimals() {
-    return this.tokenConfig.decimals;
-  }
-
-  get displayName() {
-    return this.tokenConfig.symbol;
-  }
-
-  get network() {
-    return this.wallet.network.displayName;
-  }
-
-  get ticker() {
-    return this.tokenConfig.symbol;
-  }
-
-  get balance() {
-    return this.balanceResource.current();
-  }
-
-  async transfer(to: string, amount: string) {
-    const chainId = this.wallet.network.chainId;
-    const signer = this.wallet.provider.getUncheckedSigner();
-    const contract = createUSDCContract(signer, chainId);
-
-    const transaction = await contract.transfer(to, utils.parseUnits(amount, this.decimals), {
-      gasLimit: 500000, // TODO: Use proper gasLimit value
+export class UsdcToken extends Erc20Token {
+  constructor(wallet: ReadyWallet) {
+    super(wallet, {
+      id: 'matic-network',
+      displayName: 'USDC',
+      ticker: 'USDC',
+      address: getContractAddress(ContractName.ERC20Token, wallet.network.chainId),
+      network: wallet.network.displayName,
     });
-
-    return transaction.wait();
   }
 }