+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ### vNext
 
-- 
+- Refactor Assets and activity stores to reduce copy & past
+- Remove usages of static token config everywhere, and fetch the data from contracts
+- Fix custom assets activity capturing and decimals detection
 
 ### v1.30.0